+++ src/main/java/edu/ksu/canvas/interfaces/ProgressReader.java
@@ -20,5 +20,5 @@ public interface ProgressReader extends CanvasReader<Progress, ProgressReader> {
      * @return The requested Progress object
      * @throws IOException When there is an error communicating with Canvas
      */
-    Optional<Progress> getProgress(Integer progressId) throws IOException;
+    Optional<Progress> getProgress(Long progressId) throws IOException;
 }
